Base de datos iniciada

+++ src/main/java/es/codeurjc/web/model/Producto.java
@@ -18,14 +18,14 @@ public class Producto{
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id = null;
     private String nombre;
-    private float precio;
+    private double precio;
     @Column(columnDefinition = "TEXT")
     private String descripcion;
 
     public Producto(){
         
     }
-    public Producto(String nombre, float precio, String descripcion){
+    public Producto(String nombre, double precio, String descripcion){
         this.nombre = nombre;
         this.precio = precio;
         this.descripcion = descripcion;
@@ -40,11 +40,11 @@ public class Producto{
         this.nombre = nombre;
     }
 
-    public float getPrecio() {
+    public double getPrecio() {
         return precio;
     }
 
-    public void setPrecio(float precio) {
+    public void setPrecio(double precio) {
         this.precio = precio;
     }
